ISR now use CLI() inside, stopping ISR any 4 sec.
Any ISR cycle now We issue a cli() so no more ISR will occur until  in the main someone restart ISR with sei().

This mean that now wind->flag's if inside ISR is useless, it is used only to signal to the main when some value are ready to be computed.