BUGFIX: LaCrosse introducing loop of N reading failed.
The Sensor is a mess, with these changes I do check the header, then, if
the checksum is ok, the data is valid, else I do repeat N times (defined
into default) the reading, then I accept the failed data as valid.

I think the CRC xor way is wrong. If the successive data are equal then
most probably the CRC is ok.

More work to do, using davis is one of the best.