Added autoswitch sensor in EEprom.
Now in the eeprom can be stored which kind of wind sensor we have,
avoiding auto selection if we don't want to.
A lacrosse miss-autoselection means (reported) that the wrong sensor
will be in use until a reset occur.
It is possible to change the stored value via debug 'sensor' command.