Fix modem initializations and serial bugs.
Rewrite IRQ Rx routine to be more clean, if the rx buffer is full then
restart, if the rx char is a non-valid normal char ignore it, if a CR is
the first char received then skip it.

After modem init, led blink a number of times depending on the number of
correct sentences comes from the modem. Anyway any sentence catch is ok
for consider the modem turned on.

Modem init dialog rewritten.
Speed of analyzing modem response increased to catch the "OK"
immediately follow the echo of the 1st AT command.

More dox.